refactor: console에서 쓰이던 컨트롤러 패키지 이동
